Move Dock reattach tests into focused suite
